requirements.txt
flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.19.2
reportlab==3.6.12
fillpdf
bs4
python-dotenv
requests